Mu.Adapter.ProtoBuf

Description

Just import the module and you can turn any value with a ToSchema and FromSchema from and to Protocol Buffers. Since Protocol Buffers need information about field identifiers, you need to annotate your schema using ProtoBufAnnotation.

Custom annotations

data ProtoBufAnnotation

Annotations for Protocol Buffers fields.

Constructors

ProtoBufId Nat
Numeric field identifier for normal fields

ProtoBufOneOfIds [Nat]
List of identifiers for fields which contain a union

Conversion using schemas

class ProtoBridgeTerm w sch (sch :/: sty) => IsProtoSchema w sch sty

Represents those Schemas which are supported by Protocol Buffers. Some values which can be represented as Terms cannot be so in Protocol Buffers. For example, you cannot have a list within an option.

toProtoViaSchema :: forall t f (sch :: Schema t f) a sty. (IsProtoSchema Maybe sch sty, ToSchema Maybe sch sty a) => a -> MessageBuilder

Conversion to Protocol Buffers mediated by a schema.

fromProtoViaSchema :: forall t f (sch :: Schema t f) a sty. (IsProtoSchema Maybe sch sty, FromSchema Maybe sch sty a) => Parser RawMessage a

Conversion from Protocol Buffers mediated by a schema. This function requires a RawMessage, which means that we already know that the Protocol Buffers message is well-formed. Use parseProtoViaSchema to parse directly from a ByteString.

parseProtoViaSchema :: forall sch a sty. (IsProtoSchema Maybe sch sty, FromSchema Maybe sch sty a) => ByteString -> Either ParseError a

Conversion from Protocol Buffers mediated by a schema. This function receives the ByteString directly, and parses it as part of its duty.

Conversion using registry

class FromProtoBufRegistry (ms :: Mappings Nat Schema') t

Represents Registrys for which every Schema is supported by the Protocol Buffers format.

Minimal complete definition

fromProtoBufRegistry'

fromProtoBufWithRegistry :: forall (r :: Registry) t. FromProtoBufRegistry r t => Parser RawMessage t

Conversion from Protocol Buffers by checking all the Schemas in a Registry.

As fromProtoViaSchema, this version requires an already well-formed Protocol Buffers message.

parseProtoBufWithRegistry :: forall (r :: Registry) t. FromProtoBufRegistry r t => ByteString -> Either ParseError t

Conversion from Protocol Buffers by checking all the Schemas in a Registry.

As parseProtoViaSchema, this version receives a ByteString and parses it as part of its duty.

Orphan instances

ProtoEnum Bool

Instance details

Methods

toProtoEnumMay :: Int32 -> Maybe Bool

fromProtoEnum :: Bool -> Int32

Alternative (Parser i)

Instance details

Methods

empty :: Parser i a

(<|>) :: Parser i a -> Parser i a -> Parser i a

some :: Parser i a -> Parser i [a]

many :: Parser i a -> Parser i [a]