Mu.Adapter.ProtoBuf.Via

Description

In order to interoperate with the proto3-wire library, we sometimes need an instance of Proto3WireEncoder. By using the wrappers in this module, such instances can be obtained automatically if the type can be turned into a Schema.

Documentation

newtype ViaToProtoBufTypeRef (ref :: TypeRef snm) t

Specifies that a type is turned into a Protocol Buffers message by using the schema as intermediate representation.

Constructors

ViaToProtoBufTypeRef

Fields

unViaToProtoBufTypeRef :: t

newtype ViaFromProtoBufTypeRef (ref :: TypeRef snm) t

Specifies that a type can be parsed from a Protocol Buffers message by using the schema as intermediate representation.

Constructors

ViaFromProtoBufTypeRef

Fields

unViaFromProtoBufTypeRef :: t

class FromProtoBufTypeRef (ref :: TypeRef snm) t where

Types which can be parsed from a Protocol Buffers message.

Methods

fromProtoBufTypeRef :: Proxy ref -> Parser RawMessage t

class ToProtoBufTypeRef (ref :: TypeRef snm) t where

Types which can be turned into a Protocol Buffers message.

Methods

toProtoBufTypeRef :: Proxy ref -> t -> MessageBuilder

Orphan instances

Proto3WireEncoder ()

Instance details

Methods

proto3WireEncode :: () -> MessageBuilder

proto3WireDecode :: Parser RawMessage ()