Mu.Quasi.ProtoBuf

Description

Read .proto files as a Schema. If you want to get the service definitions too, you should use GRpc instead.

Quasi-quoters for .proto files

protobuf :: String -> FilePath -> Q [Dec]

Reads a .proto file and generates a Schema with all the message types, using the name given as first argument.

Only for internal use

protobufToDecls :: String -> ProtoBuf -> Q [Dec]

Shared portion of Protocol Buffers and gRPC quasi-quoters.